Remove boolean reads print status report, clean up reset
+ Debugging code for volume issues.

+++ Arduino/VocatusController/Record.h
@@ -46,6 +46,6 @@ class Record
 	Drink _fastestDrink;
     int _endTime;
 	int _startTime;
-	int _volume;
+	float _volume;
 };
 #endif
